requirements.txt
fastapi
asyncio
nest-asyncio
pydantic
uvicorn
loguru
requests
cryptography
pymongo
kafka-python